Fix some warnings.
git-svn-id: http://eathena-project.googlecode.com/svn/trunk@15153 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/skill.c
@@ -76,7 +76,7 @@ int skill_name2id(const char* name)
 	if( name == NULL )
 		return 0;
 
-	return (int)strdb_get(skilldb_name2id, name);
+	return (int)(intptr_t)strdb_get(skilldb_name2id, name);
 }
 
 /// Maps skill ids to skill db offsets.
@@ -11611,7 +11611,7 @@ static bool skill_parse_row_skilldb(char* split[], int columns, int current)
 	skill_split_atoi(split[14],skill_db[i].blewcount);
 	safestrncpy(skill_db[i].name, trim(split[15]), sizeof(skill_db[i].name));
 	safestrncpy(skill_db[i].desc, trim(split[16]), sizeof(skill_db[i].desc));
-	strdb_put(skilldb_name2id, skill_db[i].name, (void*)id);
+	strdb_put(skilldb_name2id, skill_db[i].name, (void*)(intptr_t)id);
 
 	return true;
 }
